+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "livekit-runtime",
     "livekit-wakeword",
     "libwebrtc",
+    "libargus-sys",
     "soxr-sys",
     "yuv-sys",
     "imgproc",
@@ -48,6 +49,7 @@ license = "Apache-2.0"
 [workspace.dependencies]
 device-info = { version = "0.1.1", path = "device-info" }
 imgproc = { version = "0.3.19", path = "imgproc" }
+libargus-sys = { version = "0.1.0", path = "libargus-sys" }
 libwebrtc = { version = "0.3.40", path = "libwebrtc" }
 livekit = { version = "0.7.51", path = "livekit" }
 livekit-api = { version = "0.5.5", path = "livekit-api" }
