+++ libargus-sys/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "libargus-sys"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+description = "Native NVIDIA libargus (Jetson) capture shim and FFI bindings"
+repository.workspace = true
+links = "lk_argus"
+
+[build-dependencies]
+cc = { workspace = true }
